Move filter strategies into filter::strategy module
