+++ tiledb/api/src/array/attribute/strategy.rs
@@ -3,7 +3,7 @@ use serde_json::json;
 
 use crate::array::{attribute::FillData, AttributeData};
 use crate::datatype::strategy::*;
-use crate::filter_list::FilterListData;
+use crate::filter::list::FilterListData;
 use crate::{fn_typed, Datatype};
 
 pub fn prop_attribute_name() -> impl Strategy<Value = String> {
@@ -28,7 +28,7 @@ fn prop_fill<T: Arbitrary>() -> impl Strategy<Value = T> {
 }
 
 fn prop_filters(datatype: Datatype) -> impl Strategy<Value = FilterListData> {
-    crate::strategy::filter::prop_filter_pipeline_for_datatype(datatype)
+    crate::filter::strategy::prop_filter_pipeline_for_datatype(datatype)
 }
 
 pub fn prop_attribute_for_datatype(
