+++ tiledb/api/src/lib.rs
@@ -44,13 +44,9 @@ pub mod convert;
 pub mod datatype;
 pub mod error;
 pub mod filter;
-pub mod filter_list;
 pub mod query;
 pub mod string;
 pub mod vfs;
-
-#[cfg(feature = "proptest-strategies")]
-pub mod strategy;
 
 pub fn version() -> (i32, i32, i32) {
     let mut major: i32 = 0;
