--- tiledb/api/src/strategy/mod.rs
@@ -1,3 +1,0 @@
-pub mod filter;
-
-pub use filter::*;
